requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
accelerate
qwen-vl-utils
decord
pillow
numpy
pandas
openpack-toolkit
tqdm